refactor: Generation 모델 UMCFoundation으로 이동
- NoticeDomain에 중복 정의되어 있던 Generation struct 제거
- UMCFoundation/Generation.swift에 public init(value:) 추가
- Presentation 레이어에서 NoticeDomain.Generation이 아닌
  UMCFoundation.Generation을 참조하도록 모호성 해소
